envio de parametros al back

+++ app/simulation/components/contenedor-simulacion.tsx
@@ -5,19 +5,56 @@ import { useRouter } from 'next/navigation';
 import {
   Map,
 } from '@vis.gl/react-maplibre';
+import styles from '../../stylesheets/contenedor.module.css';
+import {FormControl, InputLabel, MenuItem, Slider} from "@mui/material";
+import {Select} from "@mui/material";
+import {useEffect, useState} from "react";
+import DatePicker from "react-datepicker";
+import "react-datepicker/dist/react-datepicker.css";
+
+enum SimulationType{
+    TIEMPO_REAL,
+    VENTANA_TIEMPO,
+    COLAPSO_OPERATIVO
+}
 
 export function ContenedorSimulacion() {
   const router = useRouter();
+  const [simulationType,setSimulationType] = useState<SimulationType>(SimulationType.COLAPSO_OPERATIVO);
+  const [startDate, setStartDate] = useState<Date|null>(new Date('2026-02-10'));
+  const [endDate, setEndDate] = useState<Date|null>(new Date('2026-02-15'));
+  const [timeScale,setTimeScale] = useState<number>(10);
+
+
 
   const handlePreparar = async () => {
+      //Validar seleccion de tipo de simulacion
+      if(simulationType==null)throw new Error("Tienes que seleccionar el tipo de simulacion");
+      //Validar existencia de fechas
+      if(startDate==null)
+          throw new Error("Tienes que seleccionar la fecha de inicio");
+      //Validar rango de fechas
+      if(simulationType===SimulationType.VENTANA_TIEMPO){
+          if(endDate == null)throw new Error("Tienes que seleccionar la fecha de fin");
+          const dateDistance = (endDate.getTime() - startDate.getTime()) / 1000 / 60 / 60 / 24 //en dias
+          if(dateDistance < 3 || dateDistance > 5) throw new Error("Rango de fechas debe de estar entre 3 y 5 dias");
+      }
+      const _startDate = startDate;
+      const _endDate =
+          (   simulationType===SimulationType.VENTANA_TIEMPO ? endDate :
+              simulationType===SimulationType.TIEMPO_REAL ? new Date(_startDate.getTime() + 1000*60*60*24) : null)
       localStorage.setItem("fechaInicio",'2026-02-10')
-    const params : ParametrosSimulacion = { fechaInicio: '2026-02-10',fechaFin: '2026-02-15' , k: 300};
+      console.log(_startDate);
+      console.log(_endDate);
+    const params : ParametrosSimulacion = { fechaInicio: _startDate.toDateString(),fechaFin: _endDate?.toDateString() , k: 300};
     const { data } = await SimulacionService.prepararInicio(params); 
     router.push(`/simulation/${data.simulacionId}?topic=${encodeURIComponent(data.websocketTopic)}`);
   };
 
+
+
   return (
-    <div style={{ position: 'relative', height: '100vh', width: '100%' }}>      
+    <div className={styles.mapContainer}>
         <Map
             initialViewState={{ longitude: 0, latitude: 0, zoom: 3.5 }}
             mapStyle="https://demotiles.maplibre.org/style.json"
@@ -25,36 +62,54 @@ export function ContenedorSimulacion() {
         >            
         </Map>
             
-      <div style={{ 
-        position: 'absolute', 
-        top: '15%', left: '50%', 
-        transform: 'translate(-50%, -50%)', 
-        zIndex: 10,
-        backgroundColor: 'rgba(30, 41, 59, 0.95)', 
-        color: '#f8fafc', 
-        padding: '30px',
-        borderRadius: '12px',
-        boxShadow: '0 20px 25px -5px rgba(0, 0, 0, 0.5)',
-        textAlign: 'center',
-        minWidth: '300px'
-        }}>
-        <h2 style={{ marginTop: 0, marginBottom: '20px' }}>Configurar Simulación</h2>
-        <p>Acá irá el formulario para configurar fecha inicio, fin y escenario, por ahora esta harcoded</p>
+      <div className={styles.simulacionModal}>
+        <h2 className={styles.simulacionTitle}>Configurar Simulación</h2>
+          <FormControl>
+              <InputLabel id="select_sim_type_label">Seleccionar tipo de simulación</InputLabel>
+              <Select id="simulation_type"
+                  labelId="select_sim_type_label"
+                  value={simulationType}
+                  onChange={(e)=>{
+                      if(typeof e.target.value === "string")return;
+                      setSimulationType(e.target.value)}}
+              >
+                  <MenuItem value={SimulationType.TIEMPO_REAL}>Ejecución en tiempo real</MenuItem>
+                  <MenuItem value={SimulationType.VENTANA_TIEMPO}>Ventana de 3 a 5 días</MenuItem>
+                  <MenuItem value={SimulationType.COLAPSO_OPERATIVO}>Hasta el colapso operativo</MenuItem>
+              </Select>
+              {simulationType === SimulationType.VENTANA_TIEMPO && (<>
+                <InputLabel id="select_date_range_label">Seleccionar rango de fechas de simulacion</InputLabel>
+                <DatePicker aria-labelledby="select_date_range_label" showMonthYearDropdown={true} selectsRange={true}
+                    startDate={startDate} endDate={endDate}
+                    selectedDates={[new Date('2026-02-10'),new Date('2026-02-15')]}
+                    onChange={(update)=> {setStartDate(update[0]);setEndDate(update[1]);}}
+                >
+                </DatePicker>
+              </>)}
+              {(simulationType === SimulationType.TIEMPO_REAL || simulationType === SimulationType.COLAPSO_OPERATIVO) && (
+                  <>
+                      <InputLabel id="select_date_start_label">Seleccionar fecha inicio</InputLabel>
+                      <DatePicker aria-labelledby="select_date_start_label" showMonthYearDropdown={true}
+                                  onChange={(update:Date|null)=> {setStartDate(update);}}
+                                  selected={startDate}
+                      >
+                      </DatePicker>
+                  </>
+              )}
+              {(simulationType === SimulationType.VENTANA_TIEMPO || simulationType === SimulationType.COLAPSO_OPERATIVO) && (
+                  <>
+                      <InputLabel id="select_time_scale_label">Numero de segundos simulados por segundo real</InputLabel>
+                      <Slider aria-labelledby="select_time_scale_label"
+                      defaultValue={10} min={1} max={99.9999} step={0.1}
+                      onChange = {(_,update)=>{setTimeScale(update);}}
+                      valueLabelDisplay="auto"
+                      />
+                  </>
+              )}
+          </FormControl>
         <button 
             onClick={handlePreparar} 
-            style={{ 
-            padding: '12px 24px', 
-            cursor: 'pointer',
-            backgroundColor: '#3b82f6',
-            color: 'white',
-            border: 'none',
-            borderRadius: '8px',
-            fontSize: '1rem',
-            fontWeight: 'bold',
-            boxShadow: '0 4px 6px -1px rgba(59, 130, 246, 0.5)'
-            }}            
-            onMouseOver={(e) => e.currentTarget.style.backgroundColor = '#2563eb'} 
-            onMouseOut={(e) => e.currentTarget.style.backgroundColor = '#3b82f6'}
+            className={styles.simulacionButton}
         >
             Preparar e Iniciar
         </button>
